better handle karpenter nodes during cleanup
Previously our cleanup operation was:

+ terraform-destroy k8s stuff
+ terraform-destroy aws stuff

However this is both inefficent and incomplete.

It's inefficient because we don't actually care about any of the stuff
running in k8s when we're shutting down.

And it's incomplete because terraform doesn't know about the resources
provisioned by Karpenter. We could delete the CRDs and wait, but I'm
lazy.

Instead we:

+ Delete all Karpenter nodepools in k8s
+ Manually invoke the AWS-cli to delete any nodes provisioned by
  Karpenter
+ terraform-destroy the rest of our AWS stuff