stop self-managing network addons
I don't love over-writing the CNI addon with our own copy of it, but the
alternative is doing "self management" of all networking plugins.

That doesn't sound bad - except the coredns addon can't be installed
until after we have nodes running which leads to awkward
order-of-operations issues now that we no longer manage nodes in TF.

We could run coredns on fargate to work-around that, but I don't miss
not having to manage version-numbers for coredns and kubeproxy.